+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2] - 2021-02-28
 
-* TODO: Describe initial release.
+* Switch from `getPurchaseHistory()` to `getAvailablePurchases()`, which seems
+to be more correct behavior on Android.
+* Add the ability to log in release mode, in case you need to debug.
+
+## [0.0.1] - initial release
+
+* First pass